Don't throw away samples
Due to discontinuity in the sweep there can be a big spike
in the amplitude at the beginning of the sweep. Now first
third of the samples are not taken into consideration in autogain
algorithm.

Reduce ADC sampling rate, because previous one was too high considering
the maximum range of the radar.